Embedded simple line graph Dash app into main Flask application
