+++ flaskPracticalWebapp/__init__.py
@@ -35,8 +35,10 @@ def create_app(config_class=Config):
         app.register_blueprint(main)
         app.register_blueprint(errors)
 
-        from flaskPracticalWebapp.plotlydash.dashboard import create_dashboard
-        app = create_dashboard(app)
+        # from flaskPracticalWebapp.plotlydash.dashboard import create_dashboard
+        # app = create_dashboard(app)
 
+        from flaskPracticalWebapp.plotlydash.dash_practical import practical_view
+        app = practical_view(app)
 
         return app
